+++ sdk_nnbd/lib/_http/http.dart
@@ -1404,6 +1404,12 @@ abstract class HttpClient {
   @Deprecated("Use defaultHttpsPort instead")
   static const int DEFAULT_HTTPS_PORT = defaultHttpsPort;
 
+  /// Enable logging of HTTP requests from all [HttpClient]s to the developer
+  /// timeline.
+  ///
+  /// Default is `false`.
+  static bool enableTimelineLogging;
+
   /// Gets and sets the idle timeout of non-active persistent (keep-alive)
   /// connections.
   ///